+++ app/src/main/java/com/example/playlistmaker/presentation/search/fragment/SearchFragment.kt
@@ -1,46 +1,25 @@
 package com.example.playlistmaker.presentation.search.fragment
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
-import android.view.inputmethod.InputMethodManager
-import android.widget.EditText
-import android.widget.ImageView
-import android.widget.ProgressBar
-import android.widget.TextView
-import androidx.appcompat.widget.AppCompatEditText
 import androidx.compose.runtime.State
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.FragmentSearchBinding
 import com.example.playlistmaker.domain.models.Track
-import com.example.playlistmaker.presentation.search.adapters.TrackAdapter
 import com.example.playlistmaker.presentation.search.listener.OnTrackClickListener
-import com.example.playlistmaker.presentation.search.viewModel.SearchScreenState
 import com.example.playlistmaker.presentation.search.viewModel.SearchViewModel
-import com.example.playlistmaker.presentation.settings.fragments.SettingsFragment
-import com.example.playlistmaker.presentation.settings.fragments.SettingsScreen
+import com.example.playlistmaker.presentation.theme.ThemeViewModel
 import com.example.playlistmaker.ui.PlaylistMakerTheme
 import com.example.playlistmaker.utils.debounce
-import com.example.playlistmaker.presentation.theme.ThemeViewModel
 import org.koin.androidx.viewmodel.ext.android.activityViewModel
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class SearchFragment : Fragment(), OnTrackClickListener {
     private val viewModel: SearchViewModel by activityViewModel()
